Transform ncurses i32 key into Event enum
Prepares support for unicode char input spanning multiple ncurses
characters.

+++ examples/lorem.rs
@@ -15,7 +15,7 @@ fn main() {
     let mut siv = Cursive::new();
 
     // We can quit by pressing q
-    siv.add_global_callback('q' as i32, |s| s.quit());
+    siv.add_global_callback('q', |s| s.quit());
 
     // The text is too long to fit on a line, so the view will wrap lines,
     // and will adapt to the terminal size.
